Chore: Move handler for onCompletion to its own file
1. Export an instance of analyzer from the source file and
use this instance through out the program
2. Move the handler function for onCompletion event to its own
file for the convenience of testing
